BSZ-229: Add and configure pytest-testmon
Since the current test suite takes up to 30 seconds to finish (mainly
due to the MCMC calculations), the `pdm run quicktest` command was
introduced to run the tests with `pytest-testmon` to only run the tests
that are covering files that have been updated.

It remains good practice to run the complete test suite using `pdm run
test` to verify *all* tests remain green.